整理代码，去除错误的注解
